✨ feat: 실시간 심박수 데이터 조회 기능 구현
- Controller: 대시보드 조회용 API 구현 (GET /api/heart-rates/{userId})
- Service: 통계, 차트(최근 30분), 리스트 데이터 통합 로직 구현
- Repository: QueryDSL을 사용한 통계(avg, max) 및 차트 데이터 조회 쿼리 작성
- DTO: HeartRateDataResponse 및 내부 레코드(Stats, Chart, List) 구조 설계

+++ lifeFit/src/main/java/com/kh/lifeFit/dto/heartData/monitoringPage/HeartDataStatsDto.java
@@ -2,7 +2,7 @@ package com.kh.lifeFit.dto.heartData.monitoringPage;
 
 public record HeartDataStatsDto(
         Integer currentHeartRate, // 현재 심박수
-        Double avgHeartRate,      // 평균 심박수 (소수점 가능)
+        Integer avgHeartRate,     // 평균 심박수 (Doubled과 int는 소수점(int는 소수점은 버려서 안됨))
         Integer maxHeartRate      // 최대 심박수
 ) {
 }
